Update docker, fix i18next loading outside tests
