+++ .dockerignore
@@ -8,6 +8,9 @@
 !/tsconfig.json
 !/build.ts
 
+# And the locales, those are important
+!/locales
+
 # Allow the tests, so they can run automatically
 # It would be better if we didn't have to build them in, but they're small, it's fine
 !/tests
