+++ locales/en/customlevel.json
@@ -23,5 +23,5 @@
   "romhackRemoved": "ROMhack has been removed as a custom level.",
   "romhackAdded": "ROMhack has been added as a custom level (enabled={{enabled}}).",
   "unclearedRemoved": "Uncleared has been removed as a custom level.",
-  "unclearedAdded": "Uncleared has been added as a custom lsadfsdevel (enabled={{enabled}})."
+  "unclearedAdded": "Uncleared has been added as a custom level (enabled={{enabled}})."
 }
